2008-02-25  Richard Hult  <richard@imendio.com>

	Merged from trunk:

	* gdk/quartz/gdkevents-quartz.c
	(_gdk_quartz_events_trigger_crossing_events): Fix build on tiger.

svn path=/branches/gtk-2-12/; revision=19645

+++ gdk/quartz/gdkevents-quartz.c
@@ -1093,7 +1093,7 @@ _gdk_quartz_events_trigger_crossing_events (void)
   gint y;
   GdkWindow *mouse_window;
   GdkWindowImplQuartz *impl;
-  NSUInteger flags = 0;
+  guint flags = 0;
   NSTimeInterval timestamp = 0;
   NSEvent *current_event;
   NSEvent *nsevent;
